pyproject.toml
[tool.poetry]
name = "fubini-visualized"
version = "0.1.0"
description = "Fubini formulas on sums visualized with Manim"
authors = ["Ewen Le Bihan <ewen.lebihan7@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
manimce = "^0.1.0"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
